+++ eva/src/components/dashBoard.vue
@@ -130,8 +130,10 @@
           <div class="settings-dash">
             <v-dialog
               ref="fullScreenModal"
-              v-model="fullScreenMode"
+              v-model="bigSizeMode"
               width="100%"
+              :fullscreen="isFullScreen"
+              style="z-index: 999"
             >
               <template v-slot:activator="{ on: onFullScreen }">
                 <v-tooltip
@@ -145,17 +147,19 @@
                       class="expand"
                       :color="theme.$main_border"
                       v-on="{ ...onFullScreen, ...onTooltip }"
-                      @click="fullScreenMode = !fullScreenMode"
+                      @click="bigSizeMode = !bigSizeMode"
                     >
                       {{ props.mdiArrowExpand }}
                     </v-icon>
                   </template>
-                  <span>На весь экран</span>
+                  <span>Развернуть</span>
                 </v-tooltip>
               </template>
               <div
                 class="full-screen-dialog"
-                :style="{ height: '80vh' }"
+                :style="{
+                  height: isFullScreen ? '100vh' : '80vh'
+                }"
               >
                 <v-card
                   class="dash-block"
@@ -259,12 +263,37 @@
                               class="expand"
                               :color="theme.$main_border"
                               v-on="on"
-                              @click="fullScreenMode = !fullScreenMode"
+                              @click="toggleBigSize"
                             >
                               {{ props.mdiArrowCollapse }}
                             </v-icon>
                           </template>
                           <span>Свернуть</span>
+                        </v-tooltip>
+                        <v-tooltip
+                          bottom
+                          :color="theme.$accent_ui_color"
+                          :open-delay="tooltipOpenDelay"
+                        >
+                          <template v-slot:activator="{ on }">
+                            <v-icon
+                              class="expand"
+                              :color="theme.$main_border"
+                              v-on="on"
+                              @click="isFullScreen = !isFullScreen"
+                            >
+                              {{
+                                isFullScreen
+                                  ? props.mdiFullscreenExit
+                                  : props.mdiFullscreen
+                              }}
+                            </v-icon>
+                          </template>
+                          <span>{{
+                            isFullScreen
+                              ? 'Выход из полноэкранного режима'
+                              : 'на весь экран'
+                          }}</span>
                         </v-tooltip>
                       </div>
                     </div>
@@ -425,7 +454,7 @@
       <v-card-text
         :is="currentElem"
         v-if="showElement"
-        :full-screen-mode="fullScreenMode"
+        :full-screen-mode="bigSizeMode"
         custom-class="card-text element-itself"
         :color-from="theme"
         :custom-style="{
@@ -440,8 +469,8 @@
         :time-format-from="props.timeFormat"
         :size-tile-from="props.sizeTile"
         :size-from="{
-          height: fullScreenMode ? fullScreenHeight : height,
-          width: fullScreenMode ? fullScreenWidth : width,
+          height: bigSizeMode ? fullScreenHeight : height,
+          width: bigSizeMode ? fullScreenWidth : width,
         }"
         :tooltip-from="props.tooltip"
         :width-from="width"
@@ -449,8 +478,8 @@
         :options="props.options"
         :current-settings="settings"
         :update-settings="updateSettings"
-        :is-full-screen="fullScreenMode"
-        :full-screen="fullScreenMode"
+        :is-full-screen="bigSizeMode"
+        :full-screen="bigSizeMode"
         :table-per-page="tablePerPage"
         :table-page="tablePage"
         :selected-pie-index="selectedPieIndex"
@@ -486,6 +515,8 @@ import {
   mdiPencil,
   mdiSettings,
   mdiTrashCanOutline,
+  mdiFullscreen,
+  mdiFullscreenExit,
 } from '@mdi/js';
 import settings from '../js/componentsSettings';
 import visualisation from '../js/visualisationCRUD';
@@ -541,7 +572,8 @@ export default {
       dataFromDB: true,
       mdiDatabaseSearch,
       mdiArrowDownBold,
-      fullScreenMode: false,
+      bigSizeMode: false,
+      isFullScreen: false,
       disabledTooltip: false,
       settings: {
         showTitle: true,
@@ -563,6 +595,8 @@ export default {
         mdiChevronDown,
         mdiArrowExpand,
         mdiArrowCollapse,
+        mdiFullscreen,
+        mdiFullscreenExit,
         icons: {},
         edit: true,
         edit_icon: true,
@@ -601,12 +635,12 @@ export default {
         tooltip: {},
         metricsMulti: [],
       },
-      fullScreenWidth: 0.8 * window.innerWidth,
-      fullScreenHeight: 0.8 * window.innerHeight,
       selectedPieIndex: -1,
       tuneValue: '',
       tuneSliderValue: '',
       nameMenu: false,
+      windowHeight: window.innerHeight,
+      windowWidth: window.innerWidth,
     };
   },
   computed: {
@@ -763,6 +797,24 @@ export default {
           .find((element) => element?.id === this.dataSourceId)?.sid
         : '';
     },
+    fullScreenHeight() {
+      if (this.bigSizeMode) {
+        if (this.isFullScreen) {
+          return this.windowHeight;
+        }
+        return this.windowHeight * 0.8;
+      }
+      return this.height;
+    },
+    fullScreenWidth() {
+      if (this.bigSizeMode) {
+        if (this.isFullScreen) {
+          return this.windowWidth;
+        }
+        return this.windowWidth * 0.8;
+      }
+      return this.height;
+    },
   },
   watch: {
     settingsIsOpened(to) {
@@ -855,8 +907,8 @@ export default {
       this.$set(this.props.tooltip, item, value);
     },
     onResize() {
-      this.fullScreenWidth = window.innerWidth * 0.8;
-      this.fullScreenHeight = window.innerHeight * 0.8;
+      this.windowWidth = window.innerWidth;
+      this.windowHeight = window.innerHeight;
     },
     updateSettings(localSettings) {
       this.settings = JSON.parse(JSON.stringify(localSettings));
@@ -1237,6 +1289,10 @@ export default {
           window.open(action.url, action.type);
           break;
       }
+    },
+    toggleBigSize() {
+      this.bigSizeMode = !this.bigSizeMode;
+      this.isFullScreen = false;
     },
   },
 };